Remove redundant #instance_eval call
Previously methods dynamically defined were using #instance_eval which
turned out to be redundant. The simplified version of what was happening
previously could be described as:

```ruby
class MyModel
  define_method(:some_name) do
    instance_eval do
      # .. logic goes here
    end
  end
end

```

In this case #instance evel is redundant as the block passed to
`#define_method` is always executed in the context of an instance
of MyModel class. Additionally, each call to `#instance_eval` allocates
a new T_NODE object.

In the code used for profiling this change reduces the number of object
allocations by 8%.