Simplify Hashifier
Previously each time new Model was built, we were creating a new
instance of an Hashifier which was then recursively iterating trough
each key, value pair to perform the same operation on nested hashes.

This created a lot of objects that were thrown away immediately.
Changing this to be a class method and simplifying implementation of
it's main method, this change decreases the number of object allocations
in the profiled code by 230K which is about 18% of all object
allocations performed.