Use symbols instead of strings for representing keys internally
Previously all keys in the input hash would be converted into strings. This
would unnecessarily allocate a lot of new String objects when
initializing a new instance of a model.

Since keys of the hash that gets passed as input are always the same,
let's use symbols instead to reduce the object allocation. This change
reduced the number of allocated objects in the profiled code by 300K.