Cache ivar name used for #instance_variable_get
Previously each access to a field would allocate a new String object for
an ivar name. It made little sense to do it as that name never changes.

Instead, let's create that string once and then cache it in the Field
object which is shared across all instances on the Model.

This change reduced the number of object allocations by 400K which
consists about 32% of all allocations done originally.